Improve @ember/object type tests and exposure
