+++ packages/@ember/object/type-tests/observers/index.test.ts
@@ -1,24 +1,16 @@
-import EmberObject from '@ember/object';
 import { removeObserver, addObserver } from '@ember/object/observers';
 
-class MyComponent extends EmberObject {
+class MyComponent {
   foo = 'bar';
 
   constructor() {
-    super();
-    this.addObserver('foo', this, 'fooDidChange');
-    this.addObserver('foo', this, this.fooDidChange);
     addObserver(this, 'foo', this, 'fooDidChange');
     addObserver(this, 'foo', this, this.fooDidChange);
-    this.removeObserver('foo', this, 'fooDidChange');
-    this.removeObserver('foo', this, this.fooDidChange);
     removeObserver(this, 'foo', this, 'fooDidChange');
     removeObserver(this, 'foo', this, this.fooDidChange);
     const lambda = () => {
       this.fooDidChange(this, 'foo');
     };
-    this.addObserver('foo', lambda);
-    this.removeObserver('foo', lambda);
     addObserver(this, 'foo', lambda);
     removeObserver(this, 'foo', lambda);
   }
@@ -28,6 +20,4 @@ class MyComponent extends EmberObject {
   }
 }
 
-const myComponent = MyComponent.create();
-myComponent.addObserver('foo', null, () => {});
-myComponent.set('foo', 'baz');
+new MyComponent();
